requirements.txt
# $ conda create --name <env> --file <this file>

# This file may be used to create an environment using:

# platform: win-64

blas==1.0  
click==8.1.3
colorama==0.4.6
flask==2.2.3
gunicorn==20.1.0
jinja2==3.1.2
numpy==1.24.2
opencv-python==4.7.0.72
pandas==2.0.1
pillow==9.2.0 
pip==23.1.2
python==3.9.16
pytorch==1.12.1
pytorch-lightning==2.0.2 
torchmetrics==0.10.3
torchsummary==1.5.1
torchvision==0.13.1
werkzeug==2.3.4 
markupsafe==2.1.2